+++ cpp/src/NVStrings_convert.cu
@@ -29,6 +29,7 @@
 #include "NVStringsImpl.h"
 #include "custring_view.cuh"
 #include "custring.cuh"
+#include "util.h"
 
 //
 int NVStrings::hash(unsigned int* results, bool todevice)
@@ -40,7 +41,7 @@ int NVStrings::hash(unsigned int* results, bool todevice)
     auto execpol = rmm::exec_policy(0);
     unsigned int* d_rtn = results;
     if( !todevice )
-        RMM_ALLOC(&d_rtn,count*sizeof(unsigned int),0);
+        d_rtn = device_alloc<unsigned int>(count,0);
 
     custring_view** d_strings = pImpl->getStringsPtr();
     thrust::for_each_n(execpol->on(0), thrust::make_counting_iterator<unsigned int>(0), count,
@@ -55,7 +56,7 @@ int NVStrings::hash(unsigned int* results, bool todevice)
     int zeros = thrust::count(execpol->on(0),d_rtn,d_rtn+count,0);
     if( !todevice )
     {
-        cudaMemcpy(results,d_rtn,sizeof(float)*count,cudaMemcpyDeviceToHost);
+        CUDA_TRY( cudaMemcpyAsync(results,d_rtn,sizeof(float)*count,cudaMemcpyDeviceToHost))
         RMM_FREE(d_rtn,0);
     }
     return (int)count-zeros;
@@ -71,7 +72,7 @@ int NVStrings::stoi(int* results, bool todevice)
     auto execpol = rmm::exec_policy(0);
     int* d_rtn = results;
     if( !todevice )
-        RMM_ALLOC(&d_rtn,count*sizeof(int),0);
+        d_rtn = device_alloc<int>(count,0);
 
     custring_view** d_strings = pImpl->getStringsPtr();
     thrust::for_each_n(execpol->on(0), thrust::make_counting_iterator<unsigned int>(0), count,
@@ -86,7 +87,7 @@ int NVStrings::stoi(int* results, bool todevice)
     int zeros = thrust::count(execpol->on(0),d_rtn,d_rtn+count,0);
     if( !todevice )
     {
-        cudaMemcpy(results,d_rtn,sizeof(int)*count,cudaMemcpyDeviceToHost);
+        CUDA_TRY( cudaMemcpyAsync(results,d_rtn,sizeof(int)*count,cudaMemcpyDeviceToHost))
         RMM_FREE(d_rtn,0);
     }
     return (int)count-zeros;
@@ -102,7 +103,7 @@ int NVStrings::stol(long* results, bool todevice)
     auto execpol = rmm::exec_policy(0);
     long* d_rtn = results;
     if( !todevice )
-        RMM_ALLOC(&d_rtn,count*sizeof(long),0);
+        d_rtn = device_alloc<long>(count,0);
 
     custring_view** d_strings = pImpl->getStringsPtr();
     thrust::for_each_n(execpol->on(0), thrust::make_counting_iterator<unsigned int>(0), count,
@@ -117,7 +118,7 @@ int NVStrings::stol(long* results, bool todevice)
     int zeros = thrust::count(execpol->on(0),d_rtn,d_rtn+count,0);
     if( !todevice )
     {
-        cudaMemcpy(results,d_rtn,sizeof(long)*count,cudaMemcpyDeviceToHost);
+        CUDA_TRY( cudaMemcpyAsync(results,d_rtn,sizeof(long)*count,cudaMemcpyDeviceToHost))
         RMM_FREE(d_rtn,0);
     }
     return (int)count-zeros;
@@ -133,7 +134,7 @@ int NVStrings::stof(float* results, bool todevice)
     auto execpol = rmm::exec_policy(0);
     float* d_rtn = results;
     if( !todevice )
-        RMM_ALLOC(&d_rtn,count*sizeof(float),0);
+        d_rtn = device_alloc<float>(count,0);
 
     custring_view** d_strings = pImpl->getStringsPtr();
     thrust::for_each_n(execpol->on(0), thrust::make_counting_iterator<unsigned int>(0), count,
@@ -155,7 +156,7 @@ int NVStrings::stof(float* results, bool todevice)
     int zeros = thrust::count(execpol->on(0),d_rtn,d_rtn+count,0);
     if( !todevice )
     {
-        cudaMemcpy(results,d_rtn,sizeof(float)*count,cudaMemcpyDeviceToHost);
+        CUDA_TRY( cudaMemcpyAsync(results,d_rtn,sizeof(float)*count,cudaMemcpyDeviceToHost))
         RMM_FREE(d_rtn,0);
     }
     return (int)count-zeros;
@@ -171,7 +172,7 @@ int NVStrings::stod(double* results, bool todevice)
     auto execpol = rmm::exec_policy(0);
     double* d_rtn = results;
     if( !todevice )
-        RMM_ALLOC(&d_rtn,count*sizeof(double),0);
+        d_rtn = device_alloc<double>(count,0);
 
     custring_view** d_strings = pImpl->getStringsPtr();
     thrust::for_each_n(execpol->on(0), thrust::make_counting_iterator<unsigned int>(0), count,
@@ -192,7 +193,7 @@ int NVStrings::stod(double* results, bool todevice)
     int zeros = thrust::count(execpol->on(0),d_rtn,d_rtn+count,0);
     if( !todevice )
     {
-        cudaMemcpy(results,d_rtn,sizeof(double)*count,cudaMemcpyDeviceToHost);
+        CUDA_TRY( cudaMemcpyAsync(results,d_rtn,sizeof(double)*count,cudaMemcpyDeviceToHost))
         RMM_FREE(d_rtn,0);
     }
     return (int)count-zeros;
@@ -208,7 +209,7 @@ int NVStrings::htoi(unsigned int* results, bool todevice)
     auto execpol = rmm::exec_policy(0);
     unsigned int* d_rtn = results;
     if( !todevice )
-        RMM_ALLOC(&d_rtn,count*sizeof(unsigned int),0);
+        d_rtn = device_alloc<unsigned int>(count,0);
 
     custring_view** d_strings = pImpl->getStringsPtr();
     thrust::for_each_n(execpol->on(0), thrust::make_counting_iterator<unsigned int>(0), count,
@@ -247,7 +248,7 @@ int NVStrings::htoi(unsigned int* results, bool todevice)
     int zeros = thrust::count(execpol->on(0),d_rtn,d_rtn+count,0);
     if( !todevice )
     {
-        cudaMemcpy(results,d_rtn,sizeof(unsigned int)*count,cudaMemcpyDeviceToHost);
+        CUDA_TRY( cudaMemcpyAsync(results,d_rtn,sizeof(unsigned int)*count,cudaMemcpyDeviceToHost))
         RMM_FREE(d_rtn,0);
     }
     return (int)count-zeros;
@@ -265,12 +266,12 @@ NVStrings* NVStrings::itos(const int* values, unsigned int count, const unsigned
     unsigned char* d_nulls = (unsigned char*)nullbitmask;
     if( !bdevmem )
     {
-        RMM_ALLOC(&d_values,count*sizeof(int),0);
-        cudaMemcpy(d_values,values,count*sizeof(int),cudaMemcpyHostToDevice);
+        d_values = device_alloc<int>(count,0);
+        CUDA_TRY( cudaMemcpyAsync(d_values,values,count*sizeof(int),cudaMemcpyHostToDevice))
         if( nullbitmask )
         {
-            RMM_ALLOC(&d_nulls,((count+7)/8)*sizeof(unsigned char),0);
-            cudaMemcpy(d_nulls,nullbitmask,((count+7)/8)*sizeof(unsigned char),cudaMemcpyHostToDevice);
+            d_nulls = device_alloc<unsigned char>(((count+7)/8),0);
+            CUDA_TRY( cudaMemcpyAsync(d_nulls,nullbitmask,((count+7)/8)*sizeof(unsigned char),cudaMemcpyHostToDevice))
         }
     }
 
@@ -323,12 +324,12 @@ NVStrings* NVStrings::ltos(const long* values, unsigned int count, const unsigne
     unsigned char* d_nulls = (unsigned char*)nullbitmask;
     if( !bdevmem )
     {
-        RMM_ALLOC(&d_values,count*sizeof(long),0);
-        cudaMemcpy(d_values,values,count*sizeof(long),cudaMemcpyHostToDevice);
+        d_values = device_alloc<long>(count,0);
+        CUDA_TRY( cudaMemcpyAsync(d_values,values,count*sizeof(long),cudaMemcpyHostToDevice))
         if( nullbitmask )
         {
-            RMM_ALLOC(&d_nulls,((count+7)/8)*sizeof(unsigned char),0);
-            cudaMemcpy(d_nulls,nullbitmask,((count+7)/8)*sizeof(unsigned char),cudaMemcpyHostToDevice);
+            d_nulls = device_alloc<unsigned char>(((count+7)/8),0);
+            CUDA_TRY( cudaMemcpyAsync(d_nulls,nullbitmask,((count+7)/8)*sizeof(unsigned char),cudaMemcpyHostToDevice))
         }
     }
 
@@ -621,12 +622,12 @@ NVStrings* NVStrings::ftos(const float* values, unsigned int count, const unsign
     unsigned char* d_nulls = (unsigned char*)nullbitmask;
     if( !bdevmem )
     {
-        RMM_ALLOC(&d_values,count*sizeof(float),0);
-        cudaMemcpy(d_values,values,count*sizeof(float),cudaMemcpyHostToDevice);
+        d_values = device_alloc<float>(count,0);
+        CUDA_TRY( cudaMemcpyAsync(d_values,values,count*sizeof(float),cudaMemcpyHostToDevice))
         if( nullbitmask )
         {
-            RMM_ALLOC(&d_nulls,((count+7)/8)*sizeof(unsigned char),0);
-            cudaMemcpy(d_nulls,nullbitmask,((count+7)/8)*sizeof(unsigned char),cudaMemcpyHostToDevice);
+            d_nulls = device_alloc<unsigned char>(((count+7)/8),0);
+            CUDA_TRY( cudaMemcpyAsync(d_nulls,nullbitmask,((count+7)/8)*sizeof(unsigned char),cudaMemcpyHostToDevice))
         }
     }
 
@@ -684,12 +685,12 @@ NVStrings* NVStrings::dtos(const double* values, unsigned int count, const unsig
     unsigned char* d_nulls = (unsigned char*)nullbitmask;
     if( !bdevmem )
     {
-        RMM_ALLOC(&d_values,count*sizeof(double),0);
-        cudaMemcpy(d_values,values,count*sizeof(double),cudaMemcpyHostToDevice);
+        d_values = device_alloc<double>(count,0);
+        CUDA_TRY( cudaMemcpyAsync(d_values,values,count*sizeof(double),cudaMemcpyHostToDevice))
         if( nullbitmask )
         {
-            RMM_ALLOC(&d_nulls,((count+7)/8)*sizeof(unsigned char),0);
-            cudaMemcpy(d_nulls,nullbitmask,((count+7)/8)*sizeof(unsigned char),cudaMemcpyHostToDevice);
+            d_nulls = device_alloc<unsigned char>(((count+7)/8),0);
+            CUDA_TRY( cudaMemcpyAsync(d_nulls,nullbitmask,((count+7)/8)*sizeof(unsigned char),cudaMemcpyHostToDevice))
         }
     }
 
@@ -744,7 +745,7 @@ int NVStrings::ip2int( unsigned int* results, bool bdevmem )
     auto execpol = rmm::exec_policy(0);
     unsigned int* d_rtn = results;
     if( !bdevmem )
-        RMM_ALLOC(&d_rtn,count*sizeof(unsigned int),0);
+        d_rtn = device_alloc<unsigned int>(count,0);
 
     custring_view** d_strings = pImpl->getStringsPtr();
     thrust::for_each_n(execpol->on(0), thrust::make_counting_iterator<unsigned int>(0), count,
@@ -782,7 +783,7 @@ int NVStrings::ip2int( unsigned int* results, bool bdevmem )
     int zeros = thrust::count(execpol->on(0),d_rtn,d_rtn+count,0);
     if( !bdevmem )
     {
-        cudaMemcpy(results,d_rtn,sizeof(unsigned int)*count,cudaMemcpyDeviceToHost);
+        CUDA_TRY( cudaMemcpyAsync(results,d_rtn,sizeof(unsigned int)*count,cudaMemcpyDeviceToHost))
         RMM_FREE(d_rtn,0);
     }
     return (int)count-zeros;
@@ -799,12 +800,12 @@ NVStrings* NVStrings::int2ip( const unsigned int* values, unsigned int count, co
     unsigned char* d_nulls = (unsigned char*)nullbitmask;
     if( !bdevmem )
     {
-        RMM_ALLOC(&d_values,count*sizeof(unsigned int),0);
-        cudaMemcpy(d_values,values,count*sizeof(unsigned int),cudaMemcpyHostToDevice);
+        d_values = device_alloc<unsigned int>(count,0);
+        CUDA_TRY( cudaMemcpyAsync(d_values,values,count*sizeof(unsigned int),cudaMemcpyHostToDevice))
         if( nullbitmask )
         {
-            RMM_ALLOC(&d_nulls,((count+7)/8)*sizeof(unsigned char),0);
-            cudaMemcpy(d_nulls,nullbitmask,((count+7)/8)*sizeof(unsigned char),cudaMemcpyHostToDevice);
+            d_nulls = device_alloc<unsigned char>(((count+7)/8),0);
+            CUDA_TRY( cudaMemcpyAsync(d_nulls,nullbitmask,((count+7)/8)*sizeof(unsigned char),cudaMemcpyHostToDevice))
         }
     }
 
@@ -887,13 +888,13 @@ int NVStrings::to_bools( bool* results, const char* true_string, bool bdevmem )
     if( true_string )
     {
         d_len = (int)strlen(true_string);
-        RMM_ALLOC(&d_true,d_len+1,0);
-        cudaMemcpy(d_true,true_string,d_len+1,cudaMemcpyHostToDevice);
+        d_true = device_alloc<char>(d_len+1,0);
+        CUDA_TRY( cudaMemcpyAsync(d_true,true_string,d_len+1,cudaMemcpyHostToDevice))
     }
     //
     bool* d_rtn = results;
     if( !bdevmem )
-        RMM_ALLOC(&d_rtn,count*sizeof(bool),0);
+        d_rtn = device_alloc<bool>(count,0);
 
     // set the values
     custring_view** d_strings = pImpl->getStringsPtr();
@@ -910,7 +911,7 @@ int NVStrings::to_bools( bool* results, const char* true_string, bool bdevmem )
     int falses = thrust::count(execpol->on(0),d_rtn,d_rtn+count,false);
     if( !bdevmem )
     {
-        cudaMemcpy(results,d_rtn,sizeof(bool)*count,cudaMemcpyDeviceToHost);
+        CUDA_TRY( cudaMemcpyAsync(results,d_rtn,sizeof(bool)*count,cudaMemcpyDeviceToHost))
         RMM_FREE(d_rtn,0);
     }
     if( d_true )
@@ -929,26 +930,24 @@ NVStrings* NVStrings::create_from_bools(const bool* values, unsigned int count, 
     NVStrings* rtn = new NVStrings(count);
 
     int d_len_true = strlen(true_string);
-    char* d_true = nullptr;
-    RMM_ALLOC(&d_true,d_len_true+1,0);
-    cudaMemcpy(d_true,true_string,d_len_true+1,cudaMemcpyHostToDevice);
+    char* d_true = device_alloc<char>(d_len_true+1,0);
+    CUDA_TRY( cudaMemcpyAsync(d_true,true_string,d_len_true+1,cudaMemcpyHostToDevice))
     int d_as_true = custring_view::alloc_size(true_string,d_len_true);
     int d_len_false = strlen(false_string);
-    char* d_false = nullptr;
-    RMM_ALLOC(&d_false,d_len_false+1,0);
-    cudaMemcpy(d_false,false_string,d_len_false+1,cudaMemcpyHostToDevice);
+    char* d_false = device_alloc<char>(d_len_false+1,0);
+    CUDA_TRY( cudaMemcpyAsync(d_false,false_string,d_len_false+1,cudaMemcpyHostToDevice))
     int d_as_false = custring_view::alloc_size(false_string,d_len_false);
 
     bool* d_values = (bool*)values;
     unsigned char* d_nulls = (unsigned char*)nullbitmask;
     if( !bdevmem )
     {
-        RMM_ALLOC(&d_values,count*sizeof(bool),0);
-        cudaMemcpy(d_values,values,count*sizeof(bool),cudaMemcpyHostToDevice);
+        d_values = device_alloc<bool>(count,0);
+        CUDA_TRY( cudaMemcpyAsync(d_values,values,count*sizeof(bool),cudaMemcpyHostToDevice))
         if( nullbitmask )
         {
-            RMM_ALLOC(&d_nulls,((count+7)/8)*sizeof(unsigned char),0);
-            cudaMemcpy(d_nulls,nullbitmask,((count+7)/8)*sizeof(unsigned char),cudaMemcpyHostToDevice);
+            d_nulls = device_alloc<unsigned char>(((count+7)/8),0);
+            CUDA_TRY( cudaMemcpyAsync(d_nulls,nullbitmask,((count+7)/8)*sizeof(unsigned char),cudaMemcpyHostToDevice))
         }
     }