+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Improvements
 
 - PR #354 Removed rmm submodule
+- PR #357 Improved error checking on memory allocates
 - PR #358 Added improved hash algorithm to reduce collisions
 
 ## Bug Fixes